feat: compound assignment operations in set to index node
