+++ test/codes/list.yak
@@ -13,3 +13,12 @@
 
 [1, 2, [3, 4]] 보여주기
 '홍길동' + [1, 2][(1 - 1) * 3 / 2] 보여주기
+
+분식 = [1, 2, '야옹', 4, 5]
+분식[0]  += 10
+분식[1] -=  2
+분식[2]*=3
+분식[3]/= 2
+분식[4] %=3
+
+분식 보여주기
