+++ test/codes/list.yak.out
@@ -9,3 +9,4 @@
 먹고 싶다 튀김
 [1, 2, [3, 4]]
 홍길동1
+[11, 0, 야옹야옹야옹, 2, 2]
